+++ packages/terra-framework-docs/src/terra-dev-site/test/pills/filter-pills/common/PillData.js
@@ -1,0 +1,34 @@
+const pillsData = [
+  {
+    label: 'asthma',
+    labelCategory: 'respiratory',
+    id: 'terra-filter-pills-asthma',
+  },
+  {
+    label: 'bronchitis',
+    labelCategory: 'respiratory',
+    id: 'terra-filter-pills-bronchitis',
+  },
+  {
+    label: 'fibro',
+    labelCategory: 'respiratory',
+    id: 'terra-filter-pills-fibro',
+  },
+  {
+    label: 'allergies',
+    labelCategory: 'skin',
+    id: 'terra-filter-pills-allergies',
+  },
+  {
+    label: 'hypertension',
+    labelCategory: 'cardio',
+    id: 'terra-filter-pills-hypertension',
+  },
+  {
+    label: 'cardiac',
+    labelCategory: 'cardio',
+    id: 'terra-filter-pills-cardiac',
+  },
+];
+
+export default pillsData;